backend/requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Dependencies
openai==1.3.8
transformers==4.36.0
torch==2.1.1

# Code Parsing
astunparse==1.6.3
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.4

# Documentation Generation
jinja2==3.1.2
markdown==3.5.1
pygments==2.17.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# File Processing
aiofiles==23.2.1